+++ CHANGELOG.rst
@@ -33,6 +33,8 @@ Fixed
   classes after filtering records.
 - Enhance the `save` method of the `torch_ecg.utils.utils_nn.CkptMixin` class:
   non-safe items in the configs are removed before saving the model.
+- Fix errors when monitor is specified while the validation dataloader is not
+  provided in the `torch_ecg.components.BaseTrainer` class.
 
 Security
 ~~~~~~~~